+++ src/packages/block/block-grid/index.ts
@@ -1,1 +1,2 @@
 export * from './workspace/index.js';
+export * from './types.js';
