+++ src/packages/block/block-type/types.ts
@@ -10,15 +10,3 @@ export interface UmbBlockTypeBase {
 	backgroundColor?: string;
 	editorSize?: UUIModalSidebarSize;
 }
-
-export interface UmbBlockGridType extends UmbBlockTypeBase {
-	columnSpanOptions: Array<number>;
-	allowAtRoot: boolean;
-	allowInAreas: boolean;
-	rowMinSpan: number;
-	rowMaxSpan: number;
-	thumbnail?: string;
-	areaGridColumns?: number;
-	areas: Array<any>;
-	groupKey: null | string;
-}
